Build functionality for operation buttons
Replaced the individual numeric button functions with a loop over
an array for parsimony and clarity.

Used the same structure for the operation buttons for the same
reasons. Within each operation button, a toggle for the equals
button is included so that the script doesn't attempt to
operate without the two necessary operands.

Global variables were moved to the top of the file, again for
parsimony, clarity, and convention.